backend/app/main.py
from fastapi import FastAPI
from fastapi.middleware.cors import CORSMiddleware
from dotenv import load_dotenv
from app.database.db import Base, engine
from app.routes import empleados, nomina, ausencias, auth

load_dotenv()

Base.metadata.create_all(bind=engine)

app = FastAPI(title="NominaColombia API", version="1.0.0")

# CORS - IMPORTANTE: Agregar antes de las rutas
app.add_middleware(
    CORSMiddleware,
    allow_origins=[
        "http://localhost:3000",
        "http://localhost:9002",
        "http://127.0.0.1:3000",
        "http://127.0.0.1:9002",
    ],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

app.include_router(auth.router)
app.include_router(empleados.router)
app.include_router(nomina.router)
app.include_router(ausencias.router)

@app.get("/health")
async def health_check():
    return {"status": "ok", "version": "1.0.0"}

if __name__ == "__main__":
    import uvicorn
    uvicorn.run("app.main:app", host="0.0.0.0", port=8000, reload=True)